# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "RecruitHub API",
    "description": "API documentation for RecruitHub",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000/api",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your bearer token in the format **Bearer {token}**"
      }
    },
    "schemas": {
      "AssignedByDto": {
        "type": "object",
        "required": ["employeeId", "name", "emailId"],
        "properties": {
          "employeeId": {
            "type": "string",
            "description": "Employee ID of the person who assigned",
            "example": "EMP002"
          },
          "name": {
            "type": "string",
            "description": "Name of the person who assigned",
            "example": "Admin User"
          },
          "emailId": {
            "type": "string",
            "format": "email",
            "description": "Email of the person who assigned",
            "example": "admin@company.com"
          }
        }
      },
      "CreateAssignmentDto": {
        "type": "object",
        "required": [
          "studentId",
          "studentName",
          "registrationNumber",
          "emailId",
          "collegeId",
          "collegeName",
          "driveId",
          "driveName",
          "panelId",
          "primaryPanelMember",
          "roundNumber",
          "assignedBy",
          "assignedTimestamp"
        ],
        "properties": {
          "assignmentId": {
            "type": "string",
            "format": "uuid",
            "description": "Unique identifier for the assignment (auto-generated if not provided)",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "studentId": {
            "type": "string",
            "format": "uuid",
            "description": "ID of the student",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "studentName": {
            "type": "string",
            "description": "Name of the student",
            "example": "John Doe"
          },
          "registrationNumber": {
            "type": "string",
            "description": "Registration number of the student",
            "example": "REG123"
          },
          "emailId": {
            "type": "string",
            "format": "email",
            "description": "Email of the student",
            "example": "john@example.com"
          },
          "collegeId": {
            "type": "string",
            "format": "uuid",
            "description": "ID of the college",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "collegeName": {
            "type": "string",
            "description": "Name of the college",
            "example": "Example College"
          },
          "driveId": {
            "type": "string",
            "format": "uuid",
            "description": "ID of the drive",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "driveName": {
            "type": "string",
            "description": "Name of the drive",
            "example": "Summer Internship Drive"
          },
          "panelId": {
            "type": "string",
            "format": "uuid",
            "description": "ID of the panel",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "primaryPanelMember": {
            "$ref": "#/components/schemas/PanelMemberDto"
          },
          "additionalPanelMembers": {
            "type": "array",
            "description": "List of additional panel members",
            "items": {
              "$ref": "#/components/schemas/PanelMemberDto"
            }
          },
          "roundNumber": {
            "type": "integer",
            "description": "Round number",
            "example": 1
          },
          "assignedBy": {
            "$ref": "#/components/schemas/AssignedByDto"
          },
          "assignedTimestamp": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when the assignment was created",
            "example": "2025-03-25T07:00:00.000Z"
          }
        }
      },
      "UpdateAssignmentDto": {
        "type": "object",
        "properties": {
          "panelId": {
            "type": "string",
            "format": "uuid",
            "description": "ID of the panel",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "primaryPanelMember": {
            "$ref": "#/components/schemas/PanelMemberDto"
          },
          "additionalPanelMembers": {
            "type": "array",
            "description": "List of additional panel members",
            "items": {
              "$ref": "#/components/schemas/PanelMemberDto"
            }
          },
          "roundNumber": {
            "type": "integer",
            "description": "Round number",
            "example": 2
          }
        }
      },
      "Assignment": {
        "type": "object",
        "properties": {
          "assignmentId": {
            "type": "string",
            "format": "uuid",
            "description": "Unique identifier for the assignment",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "studentId": {
            "type": "string",
            "format": "uuid",
            "description": "ID of the student",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "studentName": {
            "type": "string",
            "description": "Name of the student",
            "example": "John Doe"
          },
          "registrationNumber": {
            "type": "string",
            "description": "Registration number of the student",
            "example": "REG123"
          },
          "emailId": {
            "type": "string",
            "format": "email",
            "description": "Email of the student",
            "example": "john@example.com"
          },
          "collegeId": {
            "type": "string",
            "format": "uuid",
            "description": "ID of the college",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "collegeName": {
            "type": "string",
            "description": "Name of the college",
            "example": "Example College"
          },
          "driveId": {
            "type": "string",
            "format": "uuid",
            "description": "ID of the drive",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "driveName": {
            "type": "string",
            "description": "Name of the drive",
            "example": "Summer Internship Drive"
          },
          "panelId": {
            "type": "string",
            "format": "uuid",
            "description": "ID of the panel",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "primaryPanelMember": {
            "$ref": "#/components/schemas/PanelMemberDto"
          },
          "additionalPanelMembers": {
            "type": "array",
            "description": "List of additional panel members",
            "items": {
              "$ref": "#/components/schemas/PanelMemberDto"
            }
          },
          "roundNumber": {
            "type": "integer",
            "description": "Round number",
            "example": 1
          },
          "assignedBy": {
            "$ref": "#/components/schemas/AssignedByDto"
          },
          "assignedTimestamp": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when the assignment was created",
            "example": "2025-03-25T07:00:00.000Z"
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether the assignment is active",
            "example": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "The timestamp when the assignment was created",
            "example": "2025-03-25T07:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "The timestamp when the assignment was last updated",
            "example": "2025-03-25T07:00:00.000Z"
          }
        }
      },
      "PaginatedAssignmentResponse": {
        "type": "object",
        "properties": {
          "assignments": {
            "type": "array",
            "description": "List of assignments",
            "items": {
              "$ref": "#/components/schemas/Assignment"
            }
          },
          "total": {
            "type": "integer",
            "description": "Total number of assignments",
            "example": 100
          },
          "page": {
            "type": "integer",
            "description": "Current page number",
            "example": 1
          },
          "limit": {
            "type": "integer",
            "description": "Number of items per page",
            "example": 10
          }
        }
      },
      "PanelMemberDto": {
        "type": "object",
        "required": ["employeeId", "emailId", "name"],
        "properties": {
          "employeeId": {
            "type": "string",
            "description": "Employee ID of the panel member",
            "example": "EMP001"
          },
          "emailId": {
            "type": "string",
            "format": "email",
            "description": "Email ID of the panel member",
            "example": "primary@example.com"
          },
          "name": {
            "type": "string",
            "description": "Name of the panel member",
            "example": "John Doe"
          }
        }
      },
...
      "CreateCollegeDto": {
        "type": "object",
        "required": ["name", "city"],
        "properties": {
          "name": {
            "type": "string",
            "description": "The name of the college",
            "example": "Harvard University",
            "minLength": 5
          },
          "city": {
            "type": "string",
            "description": "The city where the college is located",
            "example": "Cambridge"
          }
        }
      },
      "UpdateCollegeDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "The name of the college",
            "example": "Harvard University (Updated)",
            "minLength": 5
          },
          "city": {
            "type": "string",
            "description": "The city where the college is located",
            "example": "Cambridge"
          }
        }
      },
      "CollegeResponseDto": {
        "type": "object",
        "properties": {
          "collegeId": {
            "type": "string",
            "description": "The unique identifier for the college",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "name": {
            "type": "string",
            "description": "The name of the college",
            "example": "Harvard University"
          },
          "city": {
            "type": "string",
            "description": "The city where the college is located",
            "example": "Cambridge"
          },
          "isDeleted": {
            "type": "boolean",
            "description": "Indicates whether the college has been soft deleted",
            "example": false
          },
          "createdTimestamp": {
            "type": "string",
            "format": "date-time",
            "description": "The timestamp when the college was created",
            "example": "2025-03-25T07:00:00.000Z"
          },
          "updatedTimestamp": {
            "type": "string",
            "format": "date-time",
            "description": "The timestamp when the college was last updated",
            "example": "2025-03-25T07:00:00.000Z"
          }
        }
      },
      "PaginationMetaDto": {
        "type": "object",
        "properties": {
          "totalItems": {
            "type": "integer",
            "description": "The total number of items across all pages",
            "example": 100
          },
          "itemsPerPage": {
            "type": "integer",
            "description": "The number of items per page",
            "example": 10
          },
          "currentPage": {
            "type": "integer",
            "description": "The current page number",
            "example": 1
          },
          "totalPages": {
            "type": "integer",
            "description": "The total number of pages",
            "example": 10
          }
        }
      },
      "PaginatedCollegeResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "description": "The array of college items for the current page",
            "items": {
              "$ref": "#/components/schemas/CollegeResponseDto"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PaginationMetaDto"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/assignments": {
      "post": {
        "tags": ["assignments"],
        "summary": "Create a new assignment",
        "description": "Create a new assignment with the provided details",
        "operationId": "createAssignment",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAssignmentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The assignment has been successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Assignment"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - validation error"
          },
          "401": {
            "description": "Unauthorized - invalid or missing authentication"
          }
        }
      },
      "get": {
        "tags": ["assignments"],
        "summary": "Get a paginated list of assignments",
        "description": "Get a paginated list of assignments with optional filtering",
        "operationId": "getAssignments",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number (starts from 1)",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of items per page",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "name": "studentId",
            "in": "query",
            "description": "Filter by student ID",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "panelId",
            "in": "query",
            "description": "Filter by panel ID",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "roundNumber",
            "in": "query",
            "description": "Filter by round number",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "driveId",
            "in": "query",
            "description": "Filter by drive ID",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of assignments retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedAssignmentResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - invalid or missing authentication"
          }
        }
      }
    },
    "/assignments/{assignmentId}": {
      "get": {
        "tags": ["assignments"],
        "summary": "Get an assignment by ID",
        "description": "Get detailed information about a specific assignment by its ID",
        "operationId": "getAssignmentById",
        "parameters": [
          {
            "name": "assignmentId",
            "in": "path",
            "description": "ID of the assignment to retrieve",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Assignment retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Assignment"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - assignment not found"
          },
          "401": {
            "description": "Unauthorized - invalid or missing authentication"
          }
        }
      },
      "put": {
        "tags": ["assignments"],
        "summary": "Update an assignment",
        "description": "Update an existing assignment's information",
        "operationId": "updateAssignment",
        "parameters": [
          {
            "name": "assignmentId",
            "in": "path",
            "description": "ID of the assignment to update",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAssignmentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Assignment updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Assignment"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - validation error or assignment not found"
          },
          "401": {
            "description": "Unauthorized - invalid or missing authentication"
          }
        }
      },
      "delete": {
        "tags": ["assignments"],
        "summary": "Delete an assignment",
        "description": "Remove an assignment by its ID",
        "operationId": "deleteAssignment",
        "parameters": [
          {
            "name": "assignmentId",
            "in": "path",
            "description": "ID of the assignment to delete",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Assignment removed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Assignment removed successfully"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - assignment not found"
          },
          "401": {
            "description": "Unauthorized - invalid or missing authentication"
          }
        }
      }
    },
    "/colleges": {
      "post": {
        "tags": ["colleges"],
        "summary": "Create a new college",
        "description": "Create a new college with the provided details",
        "operationId": "createCollege",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCollegeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The college has been successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollegeResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - validation error"
          },
          "401": {
            "description": "Unauthorized - invalid or missing authentication"
          }
        }
      },
      "get": {
        "tags": ["colleges"],
        "summary": "Get a paginated list of colleges",
        "description": "Get a paginated list of colleges with optional filtering by city",
        "operationId": "getColleges",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number (starts from 1)",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of items per page",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "name": "city",
            "in": "query",
            "description": "Filter by city",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of colleges retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedCollegeResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - invalid or missing authentication"
          }
        }
      }
    },
...
    "/colleges/{collegeId}": {
      "get": {
        "tags": ["colleges"],
        "summary": "Get a college by ID",
        "description": "Get detailed information about a specific college by its ID",
        "operationId": "getCollege",
        "parameters": [
          {
            "name": "collegeId",
            "in": "path",
            "description": "ID of the college to retrieve",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "College retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollegeResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - invalid or missing authentication"
          },
          "404": {
            "description": "College not found or has been deleted"
          }
        }
      },
      "put": {
        "tags": ["colleges"],
        "summary": "Update a college",
        "description": "Update an existing college's information",
        "operationId": "updateCollege",
        "parameters": [
          {
            "name": "collegeId",
            "in": "path",
            "description": "ID of the college to update",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCollegeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "College updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollegeResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - validation error"
          },
          "401": {
            "description": "Unauthorized - invalid or missing authentication"
          },
          "404": {
            "description": "College not found or has been deleted"
          }
        }
      },
      "delete": {
        "tags": ["colleges"],
        "summary": "Soft delete a college",
        "description": "Soft delete a college by its ID",
        "operationId": "deleteCollege",
        "parameters": [
          {
            "name": "collegeId",
            "in": "path",
            "description": "ID of the college to delete",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "College deleted successfully"
          },
          "401": {
            "description": "Unauthorized - invalid or missing authentication"
          },
          "404": {
            "description": "College not found or has been deleted"
          }
        }
      }
    }
  }
}